+++ CompilationLib/CompileRequest.cs
@@ -12,4 +12,5 @@ public class CompileRequest
     public string LibrariesPath { get; set; }
     public string PortCom { get;  set; }
     public bool ShouldDeploy { get; set; }
+    public bool ShouldBackup { get; set; }
 }