+++ CompilationLib/CompileResponse.cs
@@ -1,5 +1,4 @@
-﻿
-public class CompileResponse
+﻿public class CompileResponse
 {
     public List<string> Progress { get; set; } = new List<string>();
     public double ElapsedTimeInSeconds { get; internal set; }
@@ -8,6 +7,7 @@ public class CompileResponse
     public string Logs { get; internal set; }
     public bool IsSuccessful { get; set; }
     public string HashOfOptions { get; set; }
+    public string BackupFilePath { get; set; }
 
     public override string ToString()
     {